% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siqr.R
\name{generate.data}
\alias{generate.data}
\title{Data generation function for simulation and demonstration
There are three settings.}
\usage{
generate.data(
  n,
  true.theta = NULL,
  sigma = 0.1,
  setting = "setting1",
  ncopy = 1
)
}
\arguments{
\item{n}{sample size}

\item{true.theta}{true single-index coefficients,
default is c(1,1,1)/sqrt(3) for setting 1 and c(1,2)/sqrt(5) for other settings}

\item{sigma}{the standard deviation of error term}

\item{setting}{chose from three settings}

\item{ncopy}{generates multiple copies of data for Monte Carlo simulations}
}
\value{
X  predictors

Y  response variables

single.index.values single index term
}
\description{
Data generation function for simulation and demonstration
There are three settings.
}
