% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINDEXQ_fun.R
\name{index.gamma1}
\alias{index.gamma1}
\title{Main estimation function of single index quantile regression model.
a two step method.
revised version of index.gamma(), in that random sampling is implemented in its step 2}
\usage{
index.gamma1(y, X, p, gamma.inital, maxiter, tol, subsize, subrep)
}
\arguments{
\item{y}{- response vector; X - covariate matrix;}

\item{p}{- left-tail probability (quantile index), scalar}

\item{gamma.inital}{- starting value of gamma}

\item{subrep}{}
}
\value{
gamma - index direction, unit norm, first nonneg component
         flag.conv  - whether the iterations converge
}
\description{
Main estimation function of single index quantile regression model.
a two step method.
revised version of index.gamma(), in that random sampling is implemented in its step 2
}
\note{
TBD
}
