% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siqr.R
\name{lprq0}
\alias{lprq0}
\title{A supporting function that return the local polynomial regression quantile.
This estimates the quantile and its derivative at the point x.0}
\usage{
lprq0(x, y, h, tau = 0.5, x0)
}
\arguments{
\item{x}{covariate sequence;}

\item{y}{response sequence;}

\item{h}{bandwidth(scalar);}

\item{tau}{- left-tail probability}

\item{x0}{point at which the quantile is estimated}
}
\value{
x0  a scalar

fv  quantile est; dv - quantile derivative est
}
\description{
A supporting function that return the local polynomial regression quantile.
This estimates the quantile and its derivative at the point x.0
}
