% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siqr.R
\name{plot.siqr}
\alias{plot.siqr}
\title{plot function of siqr}
\usage{
\method{plot}{siqr}(x, ..., bootstrap.interval = FALSE)
}
\arguments{
\item{x}{The SIQR model object}

\item{...}{optional arguments}

\item{bootstrap.interval}{whether to calculate and plot bootstrap interval}
}
\value{
None
}
\description{
plot function of siqr
}
