% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siqr.R
\name{print.summary.siqr}
\alias{print.summary.siqr}
\title{Function to print summary}
\usage{
\method{print}{summary.siqr}(
  object,
  digits = max(5, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{siqr.object}{the single index quantile regression model object}
}
\value{
the summarized information object
}
\description{
Function to print summary
}
