% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINDEXQ_fun.R
\name{siqr}
\alias{siqr}
\title{Main estimation function of single index quantile regression model.
a two step method.}
\usage{
siqr(y, X, p, gamma.inital, maxiter, tol)
}
\arguments{
\item{y}{response vector;}

\item{X}{covariate matrix;}

\item{p}{left-tail probability (quantile index), scalar}

\item{gamma.inital}{starting value of gamma, the single index cooefficients}

\item{maxiter}{max iteration number}

\item{tol}{toleration for convergence}
}
\value{
gamma - index direction, unit norm, first nonneg component
         flag.conv  - whether the iterations converge
}
\description{
Main estimation function of single index quantile regression model.
a two step method.
}
\note{
in step 2 of the proposed algorithm, we may consider random sampling a "subsample" of size,
      say 5n, of the augmented sample(with sample size n^2);
       do it several times and take average of the estimates(need to write a sampling step, but not in this utility function
}
