% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siqr.R
\name{summary.siqr}
\alias{summary.siqr}
\title{Function to print summary}
\usage{
\method{summary}{siqr}(
  object,
  digits = max(5, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{object}{the single index quantile regression model object}

\item{digits}{controls digits in output}

\item{signif.stars}{whether show the significance stars}

\item{...}{extra arguments}
}
\value{
the summarized information object
}
\description{
Function to print summary
}
