% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siqr.R
\name{plot.siqr}
\alias{plot.siqr}
\title{plot function of siqr}
\usage{
\method{plot}{siqr}(model.obj, bootstrap.interval = FALSE)
}
\arguments{
\item{model.obj}{The SIQR model object}
}
\value{
None
}
\description{
plot function of siqr
}
