% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINDEXQ_fun.R
\name{siqr}
\alias{siqr}
\title{Main estimation function of single index quantile regression model.
a two step method.}
\usage{
siqr(
  y,
  X,
  tau = 0.5,
  beta.inital = NULL,
  maxiter = 40,
  tol = 1e-09,
  method = "own"
)
}
\arguments{
\item{y}{response vector;}

\item{X}{covariate matrix;}

\item{tau}{left-tail probability (quantile index), scalar}

\item{beta.inital}{starting value of beta, the single index coefficients}

\item{maxiter}{max iteration number}

\item{tol}{toleration for convergence}
}
\value{
beta - the fitted single index coefficients with unit norm and first component being non negative
        flag.conv  - whether the iterations converge
}
\description{
Main estimation function of single index quantile regression model.
a two step method.
}
\note{
in step 2 of the proposed algorithm, we may consider random sampling a "subsample" of size,
      say 5n, of the augmented sample(with sample size n^2);
       do it several times and take average of the estimates(need to write a sampling step, but not in this utility function
}
